Initial commit
A 2D game in which the player haracter stands centrally and defends
itself from an onslaught of enemies approaching from both sides. Enemies
can jump over the player, shoot projectiles, and attack the player
directly.

This is some starter code that introduces Screens, GameStates and
Entities as building blocks for the rest of the game.