Add Enemy Director
The enemy director will spawn enemies as well as tracking all non-hero
entities in the game. The director also tracks the player's score,
level, etc.

Also added code to support entities switching sides. This will
definitely be useful for projectiles which can be deflected in order to
injure the entity that fired/threw them, but may also have other
applications.