Hero has a Sword, HeroStates added
Adds the Sword class, which is an object carried by the hero to defend
himself.

The Hero carries a sword, and will be in charge of updating the sword's
position and state.

HeroStates correspond to the action being taken by the hero, which are
so far confined to pointing the sword in some direction. In the future,
the Hero may also use a gun and have an idle animation, as well as a
transition animation.

Add Sword