Change render to getRenderables
Instead of having each entity render itself, it now has the ability to
pass its renderable components up, which can be collected and rendered
by the GameScreen, with its single SpriteBatch. The SpriteBatch is not
yet implemented on the GameScreen.

The Hero can now get and set its own state, which will be coming to
Sword soon.