Add getRenderables() to World and EnemyDirector
There should really be an interface with getRenderables() as its
method... ah well. Yes, EnemyDirector passes its renderable objects up
to the World, which combines those with the background (eventually) and
the Hero renderables and sends all of those up to the GameScreen to be
rendered en mass by the SpriteBatch.

Although it may make more sense to have the GameScreen render the
background, and not just for the technical reasons. If a Screen
shouldn't be rendering the background, it should likely be a different
screen.