On state change Hero adjusts its shape to match
Different Hero states will involve the Hero changing its shape. Changing
the state now automatically updates the shape as well.

Since there is no stored "image in use now", the image does not need to
be updated. Instead, when the renderables are requested, the image that
corresponds to the current state will be selected and displayed.

The Sword's position should also be updated when the state changes.