GameState handles rendering, called by GameScreen
Each individual GameState handles its own rendering, taking the role
from the GameScreen. The Screen now calls the render() method on
whichever GameState is active, and the state then renders as is needed.
This means that each GameState needs to have a reference to the
SpriteBatch-- this reference should be shared no more than necessary,
and this is hopefully the most it will need to be shared.

Any GameState can call the render() method on its parent to render all
details that its parent was capable of rendering. Nesting States too
deep should be avoided, as it is a good assumption that each render()
method both begins and ends the SpriteBatch's buffer cycle, which means
that the entire screen will be redrawn for each State.