Duplicate rendering code to optimize drawing
Although premature optimization is the enemy of freedom, I still feel
bad about the horrendous inefficiency of the previous rendering for the
GameOverState. Especially with the getRenderables() method, it seems a
bit of a crime to not use it properly in drawing. This way, each region
of the screen is only drawn once.