Add InputHandler layer
As explained in the docs, an InputHandler will be specific to the type
of input being received, whether touch, keyboard, or more esoteric. The
IH then marshalls the input data and the Hero's HeroStateMediator can
query the IH in order to dictate the appropriate state for the Hero
based on the input.