Add InputProcessor to catch game-resume events
The dichotomy between the Input-polling and the InputProcessor-observer
patterns being merged doesn't thrill me at all. I could replace
everything I have now with InputProcessors in order to simplify things a
bit there. I could also have the newly added stuff put in the desktop
and html main classes, where the IP could be added to Gdx.input
directly, and have a reference to to IH in order to tell it when the
spacebar has been pressed.

In short, expect this stuff to change semi-regularly until I'm happy
with it.