Implement deflection and death detection
The can now, given a list of both the enemies and enemy projectiles in
the world, deflect incoming projectiles and kill enemies that are
colliding with the sword. Similarly, given a list of entities it can
detect whether any of them are colliding with it and should thus do
damage to the Hero.