Get base game displaying hero and sword
Lots of renaming things (sad face) in order to follow conventions. I thought I was smarter than all of that, but it seems that I am not.

There are a few changes to draw things properly with rotation as well as new positions. Development can continue as normal, hopefully with more testing. Overall, fewer crashes than I was expecting.