Hero transitions between states
Instead of going straight from one state to another based on user input, the new Hero State Controller will filter the results from the InputStateMapper (nee HeroStateMediator) and decide what is best for the user (since clearly the user can't be trusted). This functions as a way to delay user actions, making the user plan ahead a bit (0.2s currently) and increase the challenge a bit.