Textures are drawn with respect to the Polygon
Instead of drawing textures as large as they were (no larger, no smaller) and disregarding the polygonal representation of objects except in collision, textures are now drawn with the same size as their polygon representation, and in the same place.
Currently, drawing does not support polygons that do not have their bottom-left corner at (0,0). This may be fixed at some point, but it's low priority.